Added tests for Linux tag file #2852
